requirements.txt
gradio==3.47.1
openai==0.28.1
python-dotenv==1.0.0